Split AdventureManager.tsx logic to custom hooks
