+++ src/components/AdventureManager/AdventureManager.tsx
@@ -1,14 +1,4 @@
-import { useEffect, useState, useCallback } from "react";
-import { useNavigate } from "react-router-dom";
-import {
-  listStories,
-  deleteAdventure,
-  createAdventure,
-  type StoredAdventure,
-} from "@/data/adventureDatabase";
-import adventureTemplate from "@/data/adventure.yaml?raw";
-import { getAdventureTestRoute } from "@/constants/routes";
-import { importYamlFile } from "@/utils/importYaml";
+import { useCallback } from "react";
 import {
   StoriesLoadError,
   StoryCreateError,
@@ -25,97 +15,30 @@ import {
   AdventureManagerLoading,
   AdventureManagerList,
 } from "./AdventureManager.styles";
+import { useAdventureStories } from "./useAdventureStories";
+import { useAdventureImport } from "./useAdventureImport";
+import { useDeleteConfirmation } from "./useDeleteConfirmation";
 
 export const AdventureManager = () => {
-  const [stories, setStories] = useState<StoredAdventure[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<"load" | "create" | "delete" | null>(null);
-  const [deletingAdventureId, setDeletingAdventureId] = useState<string | null>(
-    null
-  );
-  const [importError, setImportError] = useState<string | null>(null);
-  const navigate = useNavigate();
+  const {
+    stories,
+    loading,
+    error,
+    handleOpenAdventure,
+    handleCreateAdventure,
+    handleDeleteAdventure,
+    loadStories,
+  } = useAdventureStories();
 
-  const loadStories = useCallback(async () => {
-    try {
-      setLoading(true);
-      setError(null);
-      const loadedStories = await listStories();
-      setStories(loadedStories);
-    } catch (err) {
-      console.error("Error loading stories:", err);
-      setError("load");
-    } finally {
-      setLoading(false);
-    }
-  }, []);
+  const { importError, handleFileImport, handleCloseImportError } =
+    useAdventureImport(loadStories);
 
-  useEffect(() => {
-    loadStories();
-  }, [loadStories]);
-
-  const handleOpenAdventure = useCallback(
-    (id: string) => {
-      navigate(getAdventureTestRoute(id));
-    },
-    [navigate]
-  );
-
-  const handleCreateAdventure = useCallback(async () => {
-    const title = "Untitled adventure";
-
-    try {
-      // Use the adventure.yaml template and replace the title
-      const contentWithNewTitle = adventureTemplate.replace(
-        /title:\s*"[^"]*"/,
-        `title: "${title}"`
-      );
-
-      const id = await createAdventure(title, contentWithNewTitle);
-      navigate(getAdventureTestRoute(id));
-    } catch (err) {
-      console.error("Error creating adventure:", err);
-      setError("create");
-    }
-  }, [navigate]);
-
-  const handleDeleteClick = useCallback((adventureId: string) => {
-    setDeletingAdventureId(adventureId);
-  }, []);
-
-  const handleConfirmDelete = useCallback(async () => {
-    if (!deletingAdventureId) return;
-
-    try {
-      await deleteAdventure(deletingAdventureId);
-      await loadStories();
-      setDeletingAdventureId(null);
-    } catch (err) {
-      console.error("Error deleting adventure:", err);
-      setError("delete");
-    }
-  }, [deletingAdventureId, loadStories]);
-
-  const handleCancelDelete = useCallback(() => {
-    setDeletingAdventureId(null);
-  }, []);
-
-  const handleFileImport = useCallback(
-    async (file: File) => {
-      const result = await importYamlFile(file);
-
-      if (result.success) {
-        // Reload stories to show the new adventure
-        await loadStories();
-        // Navigate to the imported adventure
-        navigate(getAdventureTestRoute(result.adventureId));
-      } else {
-        // Show error modal
-        setImportError(result.error);
-      }
-    },
-    [loadStories, navigate]
-  );
+  const {
+    deletingAdventureId,
+    handleDeleteClick,
+    handleConfirmDelete,
+    handleCancelDelete,
+  } = useDeleteConfirmation(handleDeleteAdventure);
 
   const handleFileDrop = useCallback(
     (file: File) => {
@@ -123,10 +46,6 @@ export const AdventureManager = () => {
     },
     [handleFileImport]
   );
-
-  const handleCloseImportError = useCallback(() => {
-    setImportError(null);
-  }, []);
 
   if (loading) {
     return (
